Separate Order and Person information
To create a tidier display, order or person information should only be shown when an order or person command is executed respectively.
Let's separate the information being displayed in ReadyBakey depending on what command type the user is requesting for.

This means CommandResult needs to have an additional input parameters to establish the command type. Thereafter, this will be used to determine what information is to be displayed.