Update UserGuide for EditOrder
User guide now consists of updated EditOrder command. Updated FAQ to note that customer's order will be tied to their ID instead of phone number.

Formatted add order command's documentation.

Might need to change the `addo` command documentation due to the difference in prefixes. Current `addo` is not the same as what is given in the UG.