Fix bug in UI
Remove select option in the UI. Not 100% complete as there is still option to select in fxml code. Result list is not changing the background as expected. Need to find a way to change it. Add borders to set as the distance between each card. This is the fastest way to get spacing between the cards, but is not consistent between the first card's top portion vs between other cards. Weird border outside supposed border as well.

Will attempt to fix spacing issue again and change the background of the listing of both orders and persons to be white.