Add sort orders feature
Let's create feature to sort orders by delivery date to improve user's ability to prioritise orders

Because ModelManager handles the output of orders that are being shown in the frontend, this is where the changes begin.
The order list that is returned out must be generalised to a ObservableList as it no longer can only be FilteredList, as it could be a SortedList.

Any method that requires a sorting command can call the new updatedSortedFilteredOrderList. Such is the case for FindIncompleteOrdersCommand.