Update Ui
Fixed issue with the response box not wrapping the text.

Let's remove the avatar from the final product. Attempted to add avatar into the response box but failed to do so due to the nature of the response box (text box). Will require a revamp of the response box in order to include the image, which is not wise as it is not time well-spent.