Edit parsing condition for details
Now, a user is able to parse in findo details in the format "quantity: item".
However, a user should only be allowed to parse details item and not quantity.

As long as the user parses "[+ve int]:" at the start of a `findo d/` command, it should not be allowed.

If this case was allowed, it could be confusing to the user:
For example the user parses `findo d/2: cake`: This could cause return values such as:
- 1: cake
- 3: cake

which can confuse the user due to differing quantity names. As a result, solely for findo d/, we prevent this from happening by checking if
`[quantity]:` is at the start of a findo d/ parse and return an error to the user