Update prefix defense
Add restrictions to the character length of address at 6-100 chars. Rationale for 6 is to ensure that users can put purely postal code if they have no time to key the full address.

Added restrictions to the name as well. It cannot go over 50 chars and below 2 chars.

Emails are restricted to 6-50 chars.

Testing of char limits should be done with ParserUtil. It is where all the exceptions are thrown and would be a better way to consolidate all exceptions.

Removed phone char limit validation in Phone to provide better error messages. Moved this to ParserUtil instead.

Will work on adding tests for these restrictions in the following milestone.